"changed migrations"

+++ database/migrations/2024_07_17_184926_create_attendees_table.php
@@ -4,27 +4,33 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class CreateAttendeesTable extends Migration
 {
     /**
      * Run the migrations.
+     *
+     * @return void
      */
-    public function up(): void
+    public function up()
     {
-        Schema::create('attendees', function (Blueprint $table) {
-            $table->id();
-            $table->integer('numbersOfAttendees');
-            $table->foreignId('event_id')->constrained()->onDelete('cascade');
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('attendees')) {
+            Schema::create('attendees', function (Blueprint $table) {
+                $table->id();
+                $table->integer('numbersOfAttendees');
+                $table->foreignId('event_id')->constrained()->onDelete('cascade');
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
      * Reverse the migrations.
+     *
+     * @return void
      */
-    public function down(): void
+    public function down()
     {
         Schema::dropIfExists('attendees');
     }
-};
+}
